split types for bundlers
